tasklink-backend/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
mysqlclient==2.2.0
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# WebSocket & Real-time
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# File handling
Pillow==10.1.0
python-magic==0.4.27

# Environment variables
python-dotenv==1.0.0
python-decouple==3.8

# Utilities
django-filter==23.5
openpyxl==3.1.2
pandas==2.1.4

# Development
django-extensions==3.2.3
ipython==8.18.1

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0